=> Ticket 1 Corrections Niveau API  refactoring du nom de certaines méthodes et nettoyage code commenté. Niveau BATCH  modification du format de la date de retrait dans le mail envoyé en auto pour les réservations.  Niveau CLIENT  modification du front pour la redirection en cas de tentative de reservation pour un livre déjà emprunté, liste de réservations et Controller en conséquence avec refactor de noms

+++ src/main/java/com/OC/p7v2api/controllers/UserController.java
@@ -45,9 +45,7 @@ public class UserController {
     private final UserService userService;
     private final UserDtoMapper userDtoMapper;
     private final UserAuthentication userAuthentication;
-    AuthenticationManager authenticationManager;
     private final BorrowDtoMapper borrowDtoMapper;
-    private final BorrowService borrowService;
     private final ReservationDtoMapper reservationDtoMapper;
     private final ReservationService reservationService;
     private final BookService bookService;
@@ -58,7 +56,6 @@ public class UserController {
         log.info("in UserController in AUTHENTICATE with username {}", username);
         String token = userAuthentication.successfulAuthentication(username, password);
         return new ResponseEntity(token, HttpStatus.OK);
-
     }
 
 
@@ -99,9 +96,14 @@ public class UserController {
     }
 
     @PostMapping("/users/account/reservations/delete")
-    public void deleteAReservation(@RequestParam Integer reservationId) {
+    public ResponseEntity deleteAReservation(@RequestParam Integer reservationId) {
         log.info("HTTP POST request received at /users/account/reservations/delete with borrowList where id is {} ", reservationId);
+        if (reservationId == null) {
+            log.info("HTTP POST request received at /users/account/reservations/delete where id is null");
+            return new ResponseEntity<>(reservationId, HttpStatus.NO_CONTENT);
+        }
         reservationService.deleteAReservationById(reservationId);
+        return ResponseEntity.status(HttpStatus.ACCEPTED).build();
     }
 
     @Transactional
